[wip] New job API
Before this change, we only had smart_proxy_dynflow's APIs for starting
jobs. The provided APIs were RPC-esque the caller had to know the
internals of the callee and the same thing applied when getting outputs
from completed jobs.

This set of changes formalizes jobs as something that can be created and
interacted with. A job represents run of a script against a single host.

Bulk API allows creation of multiple jobs with a single request. A bulk
is a purely abstract concept, the jobs created in bulk are completely
separate and know nothing of their siblings within the same bulk.

The goal of these changes is introducing a sane API for manging jobs and
decoupling the API from the actual implementation.